Fix tox environment for package version problems
This is an attempt to fix package version problem caused by Routes. In
the gobal requirements (at the moment), we have to use version 1.13 of
the Routes package. However, that version is not py34 friendly. This
patch adds a separate testenv and a separate requirements-py3.txt file
to deal with this.

Change-Id: Id751f3303b546ac066f65c16385232dcc10bd790