Correct the installation doc
The instacllation doc should provide the senlin dashboard install
in devstack method, because horizon is a core service, many
developer will install this service. And if the developer want
to develop senlin-dashboard there is no doc instruction to guide
them how to install this horizon plugin.
Besides, there is a incorrect step in the install guide.

Change-Id: I375f04cb1e2621ce59740770566692e6643add05