Fix cluster status setting when RESIZE
With this patch, we are avoiding setting cluster status based on the
action (RESIZE) executed on the cluster. The cluster's status is only
set according to its member node status.

This patch also adjusts the cluster's desired_capacity before doing
resize so that 'desired_capacity' always expresses the user wanted
number.

Change-Id: Ib900e1bb05d0e9fc12efc3acb109079242b36407